+++ lib/elixir/lib/behaviour.ex
@@ -1,0 +1,52 @@
+defmodule Behaviour do
+  @moduledoc """
+  A convenient module for defining behaviors.
+  It provides a `defcallback` macro for defining
+  the callbacks and automatically generates a
+  `behaviour_info` function before compilation.
+
+      defmodule URI.Parser do
+        use Behaviour
+
+        @doc "Parses the given URL"
+        defcallback parse(arg)
+
+        @doc "Defines a default port"
+        defcallback default_port(arg)
+      end
+
+  """
+
+  @doc """
+  Annotates the given function is a callback. `defcallback` is
+  slightly different than simple using `def` because, even if
+  `defcallback` contains default values, a default function
+  won't be generated, which would happen with `def`.
+  """
+  defmacro defcallback(fun) do
+    { name, args } = Erlang.elixir_clauses.extract_args(fun)
+
+    quote do
+      @__behaviour_callbacks { unquote(name), unquote(length(args)) }
+      def unquote(fun), :skip_definition
+    end
+  end
+
+  @doc false
+  defmacro __using__(_) do
+    quote do
+      Module.register_attribute(__MODULE__, :__behaviour_callbacks, accumulate: true)
+      @before_compile unquote(__MODULE__)
+      import unquote(__MODULE__), only: [defcallback: 1]
+    end
+  end
+
+  @doc false
+  defmacro before_compile(_) do
+    quote do
+      def behaviour_info(:callbacks) do
+        @__behaviour_callbacks
+      end
+    end
+  end
+end