+++ lib/elixir/lib/gen_server/behaviour.ex
@@ -1,4 +1,15 @@
 defmodule GenServer.Behavior do
+  @moduledoc false
+
+  defmacro __using__(_) do
+    IO.puts "Using GenServer.Behavior is deprecated, use GenServer.Behaviour instead"
+    quote do
+      use GenServer.Behavior
+    end
+  end
+end
+
+defmodule GenServer.Behaviour do
   @doc """
   By using this module, you get default GenServer callbacks
   for `handle_call`, `handle_info`, `handle_cast`, `terminate`